make delete button into individual class.
